feat: buffer signal channel
This looks like a false positive in go vet. See https://github.com/golang/go/issues/45043
However, it should be fine to just buffer the channel to silence go vet,
as was discussed in the issue.

Ideally we could silence the error, but there is no way to ignore a line in a file for go vet: https://github.com/golang/go/issues/17058